LBaaS reconciliation: split into more files
Extract Object-type-specific code into extra files.
